create better logging and bmarks
