+++ deepdow/benchmarks.py
@@ -5,7 +5,7 @@ import cvxpy as cp
 from cvxpylayers.torch import CvxpyLayer
 import torch
 
-from deepdow.nn import CovarianceMatrix
+from .layers import CovarianceMatrix
 
 
 class Benchmark(ABC):
@@ -24,78 +24,58 @@ class Benchmark(ABC):
     def lookback_invariant(self):
         """Determine whether model yields the same weights irrespective of the lookback size."""
 
-    def fit(self, *args, **kwargs):
-        """Fitting of the model. By default does nothing."""
-        return self
+    @property
+    @abstractmethod
+    def deterministic(self):
+        """Determine whether model yields the same weights for the same input."""
 
 
 class MaximumReturn(Benchmark):
     """Markowitz portfolio optimization - maximum return."""
 
-    def __init__(self, max_weight=1):
+    def __init__(self, max_weight=1, n_assets=None, returns_channel=0):
         """Construct.
 
         Parameters
         ----------
         max_weight : float
             A number in (0, 1] representing the maximum weight per asset.
+
+        n_assets : None or int
+            If specifed the benchmark will always have to be provided with `n_assets` of assets in the `__call__`.
+            This way one can achieve major speedups since the optimization problem is canonicalized only once in the
+            constructor. However, when `n_assets` is None the optimization problem is canonicalized before each
+            inside of `__call__` which results in overhead but allows for variable number of assets.
+
+        returns_channel : int
+            Which channel in the `X` feature matrix to consider (the 2nd dimension) as returns.
         """
         self.max_weight = max_weight
+        self.n_assets = n_assets
+        self.return_channel = returns_channel
+
+        self.optlayer = self._construct_problem(n_assets, max_weight) if self.n_assets is not None else None
 
     @property
     def lookback_invariant(self):
         return False
 
-    def __call__(self, X):
-        """Predict weights.
+    @property
+    def deterministic(self):
+        return True
 
-        Parameters
-        ----------
-        X : torch.Tensor
-            Tensor of shape `(n_samples, 1, lookback, n_assets)`.
-
-        Returns
-        -------
-        weights : torch.Tensor
-            Tensor of shape `(n_samples, n_assets)` representing the predicted weights.
-
-        """
-        n_samples, _, lookback, n_assets = X.shape
-
-        # Problem setup
+    @staticmethod
+    def _construct_problem(n_assets, max_weight):
+        """Construct cvxpylayers problem."""
         rets = cp.Parameter(n_assets)
         w = cp.Variable(n_assets)
 
         ret = rets @ w
         prob = cp.Problem(cp.Maximize(ret), [cp.sum(w) == 1,
                                              w >= 0,
-                                             w <= self.max_weight])
+                                             w <= max_weight])
 
-        cvxpylayer = CvxpyLayer(prob, parameters=[rets], variables=[w])
-
-        # problem solver
-        rets_estimate = X[:, 0, :, :].mean(dim=1)
-        results_list = [cvxpylayer(rets_estimate[i])[0] for i in range(n_samples)]
-
-        return torch.stack(results_list, dim=0)
-
-
-class MinimumVariance(Benchmark):
-    """Markowitz portfolio optimization - minimum variance."""
-
-    def __init__(self, max_weight=1):
-        """Construct.
-
-        Parameters
-        ----------
-        max_weight : float
-            A number in (0, 1] representing the maximum weight per asset.
-        """
-        self.max_weight = max_weight
-
-    @property
-    def lookback_invariant(self):
-        return False
+        return CvxpyLayer(prob, parameters=[rets], variables=[w])
 
     def __call__(self, X):
         """Predict weights.
@@ -103,7 +83,7 @@ class MinimumVariance(Benchmark):
         Parameters
         ----------
         X : torch.Tensor
-            Tensor of shape `(n_samples, 1, lookback, n_assets)`.
+            Tensor of shape `(n_samples, n_input_channels, lookback, n_assets)`.
 
         Returns
         -------
@@ -114,21 +94,86 @@ class MinimumVariance(Benchmark):
         n_samples, _, lookback, n_assets = X.shape
 
         # Problem setup
+        if self.optlayer is not None:
+            if self.n_assets != n_assets:
+                raise ValueError('Incorrect number of assets: {}, expected: {}'.format(n_assets, self.n_assets))
+
+            optlayer = self.optlayer
+        else:
+            optlayer = self._construct_problem(n_assets, self.max_weight)
+
+        rets_estimate = X[:, self.return_channel, :, :].mean(dim=1)  # (n_samples, n_assets)
+
+        return optlayer(rets_estimate)[0]
+
+
+class MinimumVariance(Benchmark):
+    """Markowitz portfolio optimization - minimum variance."""
+
+    def __init__(self, max_weight=1, returns_channel=0, n_assets=None):
+        """Construct.
+
+        Parameters
+        ----------
+        max_weight : float
+            A number in (0, 1] representing the maximum weight per asset.
+        """
+        self.n_assets = n_assets
+        self.return_channel = returns_channel
+        self.max_weight = max_weight
+
+        self.optlayer = self._construct_problem(n_assets, max_weight) if self.n_assets is not None else None
+
+    @property
+    def lookback_invariant(self):
+        return False
+
+    @property
+    def deterministic(self):
+        return True
+
+    @staticmethod
+    def _construct_problem(n_assets, max_weight):
+        """Construct cvxpylayers problem."""
         covmat_sqrt = cp.Parameter((n_assets, n_assets))
         w = cp.Variable(n_assets)
 
         risk = cp.sum_squares(covmat_sqrt @ w)
         prob = cp.Problem(cp.Minimize(risk), [cp.sum(w) == 1,
                                               w >= 0,
-                                              w <= self.max_weight])
+                                              w <= max_weight])
 
-        cvxpylayer = CvxpyLayer(prob, parameters=[covmat_sqrt], variables=[w])
+        return CvxpyLayer(prob, parameters=[covmat_sqrt], variables=[w])
+
+    def __call__(self, X):
+        """Predict weights.
+
+        Parameters
+        ----------
+        X : torch.Tensor
+            Tensor of shape `(n_samples, n_input_channels, lookback, n_assets)`.
+
+        Returns
+        -------
+        weights : torch.Tensor
+            Tensor of shape `(n_samples, n_assets)` representing the predicted weights.
+
+        """
+        n_samples, _, lookback, n_assets = X.shape
+
+        # Problem setup
+        if self.optlayer is not None:
+            if self.n_assets != n_assets:
+                raise ValueError('Incorrect number of assets: {}, expected: {}'.format(n_assets, self.n_assets))
+
+            optlayer = self.optlayer
+        else:
+            optlayer = self._construct_problem(n_assets, self.max_weight)
 
         # problem solver
-        covmat_sqrt_estimates = CovarianceMatrix(sqrt=True)(X[:, 0, :, :])
-        results_list = [cvxpylayer(covmat_sqrt_estimates[i])[0] for i in range(n_samples)]
+        covmat_sqrt_estimates = CovarianceMatrix(sqrt=True)(X[:, self.return_channel, :, :])
 
-        return torch.stack(results_list, dim=0)
+        return optlayer(covmat_sqrt_estimates)[0]
 
 
 class OneOverN(Benchmark):
@@ -140,7 +185,7 @@ class OneOverN(Benchmark):
         Parameters
         ----------
         X : torch.Tensor
-            Tensor of shape `(n_samples, 1, lookback, n_assets)`.
+            Tensor of shape `(n_samples, n_input_channels, lookback, n_assets)`.
 
         Returns
         -------
@@ -150,10 +195,14 @@ class OneOverN(Benchmark):
         """
         n_samples, n_channels, lookback, n_assets = X.shape
 
-        return torch.ones((n_samples, n_assets), dtype=X.dtype) / n_assets
+        return torch.ones((n_samples, n_assets), dtype=X.dtype, device=X.device) / n_assets
 
     @property
     def lookback_invariant(self):
+        return True
+
+    @property
+    def deterministic(self):
         return True
 
 
@@ -166,7 +215,7 @@ class Random(Benchmark):
         Parameters
         ----------
         X : torch.Tensor
-            Tensor of shape `(n_samples, 1, lookback, n_assets)`.
+            Tensor of shape `(n_samples, n_input_channels, lookback, n_assets)`.
 
         Returns
         -------
@@ -176,7 +225,7 @@ class Random(Benchmark):
         """
         n_samples, n_channels, lookback, n_assets = X.shape
 
-        weights_unscaled = torch.rand((n_samples, n_assets), dtype=X.dtype)
+        weights_unscaled = torch.rand((n_samples, n_assets), dtype=X.dtype, device=X.device)
         weights_sums = weights_unscaled.sum(dim=1, keepdim=True).repeat(1, n_assets)
 
         return weights_unscaled / weights_sums
@@ -184,6 +233,10 @@ class Random(Benchmark):
     @property
     def lookback_invariant(self):
         return True
+
+    @property
+    def deterministic(self):
+        return False
 
 
 class Singleton(Benchmark):
@@ -218,11 +271,15 @@ class Singleton(Benchmark):
         if self.asset_ix not in set(range(n_assets)):
             raise IndexError('The selected asset index is out of range.')
 
-        weights = torch.zeros((n_samples, n_assets), dtype=X.dtype)
+        weights = torch.zeros((n_samples, n_assets), dtype=X.dtype, device=X.device)
         weights[:, self.asset_ix] = 1
 
         return weights
 
     @property
     def lookback_invariant(self):
+        return True
+
+    @property
+    def deterministic(self):
         return True
